requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
gunicorn==23.0.0
requests>=2.31.0
dnspython>=2.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
